feat: moved components into folders
