+++ src/app/components/cards/PostCard.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { PostProps } from "../../../types/types";
+import { PostProps } from "../../../../types/types";
 import Image from "next/image";
 import comment from "../../../public/images/comment.svg";
 import avatar from "../../../public/images/profile.png";
